refactor: reorganize frontend pages
