+++ frontend/src/components/Ca3Section.jsx
