+++ frontend/src/pages/CommunityPage.jsx
