+++ frontend/src/pages/HomePage.jsx
@@ -3,7 +3,7 @@ import HeroSection from '../components/HeroSection';
 import InfoButtons from '../components/InfoButtons';
 import NutritionEvaluation from '../components/NutritionEvaluation';
 import Functionalities from '../components/Functionalities';
-import CommunityPosts from '../components/CommunityPosts(unused)';
+import CommunityPosts from '../components/CommunityPosts';
 import ContactSection from '../components/ContactSection';
 
 const HomePage = () => {
