+++ frontend/src/pages/SearchResults.jsx
